Refactor: Component 구분
1. ConsumedLimitView Component구분
 - ConsumedInput: 입력페이지
 - ConsumedButton: 버튼페이지

2. ComsumedLimit관련 Component 주석 작성
