+++ Moasa/Moasa/View/InitSetting/ConsumedLimitView.swift
@@ -35,139 +35,22 @@ struct ConsumedLimitView: View {
                     .padding(.bottom, 40)
                 Spacer()
             }.padding(.leading, 16)
-            if showList {
-                Section(header: HStack {
-                    Text("소비 예산을 설정해주세요")
-                    Spacer()
-                    Button(action: {
-                        self.addArray += 1
-                        addElement()
-                        consumedLimitStr.append("")
-                        consumedLimitValues.append(0)
-                        hideKeyboard()
-                    }, label: {
-                        Image(systemName: "plus")
-                    })
-                }
-                    .font(.system(size: 20, weight: .bold))
-                    .padding(.horizontal, 16), content: {
-                        ScrollView {
-                            VStack {
-                                if addArray > 0 {
-                                    ForEach(0..<consumedCategories.count, id: \.self) { idx in
-                                        HStack {
-                                            VStack {
-                                                TextField("기타", text: $consumedCategories[idx].consumedCategory)
-                                                    .font(.system(size: 17, weight: .regular))
-                                                    .lineSpacing(0)
-                                                Divider()
-                                                    .background(Color.accentColor)
-                                            }
-                                            VStack {
-                                                TextField("0", text: $consumedLimitStr[idx])
-                                                    .onChange(of: consumedLimitStr[idx]) {newValue in
-                                                        consumedLimitValues[idx] = Int(newValue)!
-                                                    }
-                                                    .keyboardType(.numberPad)
-                                                    .padding(.trailing, 16)
-
-//                                                TextField("100000", value: $consumedLimitValues[idx],
-//                                                          formatter: NumberFormatter())
-//                                                    .keyboardType(.numberPad)
-//                                                    .padding(.trailing, 16)
-                                                Divider()
-                                                    .background(Color.accentColor)
-                                            }
-                                            Text("원")
-                                                .font(.system(size: 17, weight: .regular))
-                                        }
-                                    }
-                                    .onDelete(perform: delete)
-                                }
-                            }
-                            .padding(.horizontal, 16)
-                            .padding(.bottom, 40)
-                        }
-                    })
-            }
-            HStack {
-                Text("목표를 위해 매 월 얼마까지 모을 수 있나요?")
-                    .font(.system(size: 20, weight: .bold))
-                Spacer()
-            }.padding(.leading, 16)
-
-            HStack {
-                TextField("월 저축액을 입력해주세요.", text: $text)
-                    .onChange(of: text) {newValue in
-                        fixedSaving = Int(newValue)!
-                    }
-                    .padding(.leading, 16)
-                    .font(.system(size: 17, weight: .regular))
-                    .keyboardType(.numberPad)
-                Text("원")
-                    .font(.system(size: 17, weight: .regular))
-                    .padding(.trailing, 16)
-            }
-            Divider()
-                .background(Color.accentColor)
-                .padding(.horizontal, 16)
-                .padding(.bottom, 20)
-            Spacer()
-
-            VStack {
-                if addArray > 2 && showList {
-                    NavigationLink(destination: MainView().environmentObject(items), tag: true, selection: $nextView) {
-                        EmptyView()
-                    }
-
-                    Button(action: {
-                        // 1. 입력된 카텓고리의 모든 이름이 빈 문자열이 아니다.
-                        // 2. 입력된 카테고리의 모든 가격이 0 초과다.
-                        let categoryNameCheck = consumedCategories.filter { $0.consumedCategory.isEmpty }
-                        let consumedLimitCheck = consumedLimitValues.filter { $0 <= 0 }
-
-                        if categoryNameCheck.isEmpty && consumedLimitCheck.isEmpty {
-                            withAnimation {
-                                for idx in 0..<consumedCategories.count {
-                                    consumedCategories[idx].consumedLimit[0] = consumedLimitValues[idx]
-                                }
-                                items.consumedCategories = consumedCategories
-                                print(consumedCategories[0].consumedLimit)
-                                print(consumedCategories[1].consumedLimit)
-                                UserDefaults.standard.set(true, forKey: "initSetting")
-                            }
-                            nextView = true
-                        } else {
-                            // ALERT "이상한 정보입니다."
-                        }
-                    }, label: {
-                        BtnShape(btnText: $btnText[1])
-                    })
-                } else if self.fixedSaving >= 0 {
-                    // 예산항목 추가를 보여주는 버튼
-                    Button(action: {
-                        showList = true
-                        hideKeyboard()
-                    }, label: {
-                        BtnShape(btnText: $btnText[0])
-                    })
-                    .opacity(self.fixedSaving > 0 ? 1: 0)
-                }
-            }
+// Component - 소비 항목 입력 컴포넌트
+            ConsumedInput(consumedCategories: $consumedCategories, consumedLimitValues: $consumedLimitValues,
+                          showList: $showList, consumedLimitStr: $consumedLimitStr,
+                          addArray: $addArray, text: $text,
+                          fixedSaving: $fixedSaving)
+            .environmentObject(items)
+// Component - 저장, 네비게이션 버튼
+            ConsumedButton(consumedCategories: $consumedCategories, consumedLimitValues: $consumedLimitValues,
+                           addArray: $addArray, showList: $showList, nextView: $nextView,
+                           fixedSaving: $fixedSaving, btnText: $btnText)
+            .environmentObject(items)
         }
         .background(Color.kenCustomOrange)
         .navigationBarHidden(true)
         .onTapGesture {
             hideKeyboard()
-        }
-    }
-    func addElement() {
-        consumedCategories.append(ConsumedCategory(consumedCategory: "", consumedLimit: [:]))
-    }
-    func delete(offset: IndexSet) {
-        withAnimation {
-            consumedCategories.remove(atOffsets: offset)
-            consumedLimitValues.remove(atOffsets: offset)
         }
     }
 }
